JSON simplification: "message" result
We already had a simplification where any result object without a `type`
property would be treated as an effect. Now we tweak that and do a bit
of type inference.

If the `type` property is missing, we first check for a `message`
property, and if present, we treat the result as a MessageResult. We
then check for `name` and, if present, interpret as an EffectResult. If
both of those checks fail, we create a MessageResult with an error
message.