Add the concept of event and status templates
Events and status effects can now use another one as a template! (Events
can only use other events, and statuses can only use other statuses.)

Modified the DC for the saving throw in accident checks to scale with
the bathroom need. (Need / 10) + 10. The need for the accident check
gate to happen first may go away, or just be turned into a modifier for
the DC (e.g. if the event says you wet at 300%, the DC modifier would be
low, because it takes longer for the DC to be unbeatable).